Added support for kotlin power assert
