+++ kotest-assertions/kotest-assertions-shared/build.gradle.kts
@@ -1,7 +1,10 @@
+import org.jetbrains.kotlin.gradle.ExperimentalKotlinGradlePluginApi
+
 plugins {
    id("kotest-multiplatform-library-conventions")
    id("kotest-android-native-conventions")
    id("kotest-watchos-device-conventions")
+   kotlin("plugin.power-assert") version "2.0.0"
 }
 
 kotlin {
@@ -32,4 +35,9 @@ kotlin {
          }
       }
    }
+}
+
+@OptIn(ExperimentalKotlinGradlePluginApi::class)
+powerAssert {
+   functions = listOf("io.kotest.matchers.shouldBe")
 }
